requirements.txt
PyQt6==6.3.1
Pillow==10.2.0
reportlab==4.1.0
pytest==8.0.0
pyinstaller==6.12.0 